timeout() should take a long as a parameter. Generally libraries handle timeout milliseconds as longs for example TimeUnit.SECONDS.toMillis(1) returns a longs as well. I have to cast this value down to int every time I plan to use TimeUnit.

+++ src/org/mockito/Mockito.java
@@ -2091,7 +2091,7 @@ public class Mockito extends Matchers {
      * 
      * @return verification mode
      */
-    public static VerificationWithTimeout timeout(int millis) {
+    public static VerificationWithTimeout timeout(long millis) {
         return new Timeout(millis, VerificationModeFactory.times(1));
     }
